+++ retail/templates/serializers.py
@@ -7,7 +7,6 @@ class CreateTemplateSerializer(serializers.Serializer):
     template_translation = serializers.JSONField(required=True)
     template_name = serializers.CharField(required=True)
     category = serializers.CharField(required=True)
-    start_condition = serializers.CharField(required=True)
     app_uuid = serializers.CharField(required=True)
     project_uuid = serializers.CharField(required=True)
     rule_code = serializers.CharField(required=False)
@@ -16,15 +15,23 @@ class CreateTemplateSerializer(serializers.Serializer):
 class ReadTemplateSerializer(serializers.Serializer):
     uuid = serializers.UUIDField()
     name = serializers.CharField()
-    start_condition = serializers.CharField()
+    display_name = serializers.SerializerMethodField()
+    start_condition = serializers.SerializerMethodField()
     status = serializers.SerializerMethodField()
     rule_code = serializers.CharField()
+    metadata = serializers.JSONField()
 
     def get_status(self, obj: Template) -> str:
         if obj.current_version is not None:
             return obj.current_version.status
 
         return "PENDING"
+
+    def get_display_name(self, obj):
+        return obj.parent.display_name
+
+    def get_start_condition(self, obj):
+        return obj.parent.start_condition
 
 
 class UpdateTemplateSerializer(serializers.Serializer):